refactor: Modularize Tetris components
TBS-96
