--- components/index.tsx
@@ -1,4 +1,0 @@
-export { ControlPrompt } from "./prompt/ControlPrompt";
-export { CountdownPrompt } from "./prompt/CountdownPrompt";
-export { StartPrompt } from "./prompt/StartPrompt";
-export { Tetris } from "./game/Tetris";
