+++ styles/components/Tetris.module.css
@@ -1,119 +1,3 @@
-.info {
-  width: 280px;
-  color: #f3f4f6;
-  font-size: 1.5em;
-  font-weight: 700;
-  opacity: 0;
-  text-align: center;
-
-  @apply grid grid-cols-2 gap-x-3;
-}
-
-.transform-info {
-  animation-duration: 1.75s;
-  animation-fill-mode: forwards;
-  animation-name: info-fade-in;
-  animation-timing-function: ease-out;
-}
-
-@keyframes info-fade-in {
-  0% {
-    opacity: 0;
-  }
-  75% {
-    opacity: 0;
-  }
-  100% {
-    opacity: 1;
-  }
-}
-
-.queue {
-  @apply grid grid-rows-6 gap-1;
-}
-
-.held,
-.next {
-  min-height: 60px;
-  min-width: 60px;
-}
-
-.held {
-  opacity: 0;
-  place-self: start;
-}
-
-.transform-held {
-  animation-duration: 1.75s;
-  animation-fill-mode: forwards;
-  animation-name: held-fade-in;
-  animation-timing-function: ease-out;
-}
-
-@keyframes held-fade-in {
-  0% {
-    opacity: 0;
-  }
-  75% {
-    opacity: 0;
-  }
-  100% {
-    border: 5px solid #f3f4f6;
-    border-radius: 0.25rem;
-    opacity: 1;
-  }
-}
-
-.game {
-  border: 5px solid #f3f4f6;
-  border-radius: 0.25rem;
-  height: 440px;
-  width: 220px;
-  opacity: 0;
-}
-
-.transform-game {
-  animation-duration: 1.75s;
-  animation-fill-mode: forwards;
-  animation-name: game-fade-in;
-  animation-timing-function: ease-out;
-}
-
-@keyframes game-fade-in {
-  0% {
-    opacity: 0;
-  }
-  75% {
-    opacity: 0;
-  }
-  100% {
-    border: 5px solid #f3f4f6;
-    border-radius: 0.25rem;
-    opacity: 1;
-  }
-}
-
-.transform-queue > .next {
-  animation-duration: 1.75s;
-  animation-fill-mode: forwards;
-  animation-name: queue-fade-in;
-  animation-timing-function: ease-out;
-}
-
-@keyframes queue-fade-in {
-  0% {
-    opacity: 0;
-  }
-  75% {
-    opacity: 0;
-  }
-  100% {
-    border: 5px solid #f3f4f6;
-    border-radius: 0.25rem;
-    opacity: 1;
-  }
-}
-
 /* Styles for default tetrominos */
 .row {
   border: 1px solid transparent;
@@ -121,7 +5,6 @@
 }
 
 .row-queue {
-  border: 1px solid transparent;
   padding: 5px;
 }
