+++ styles/components/game/Held.module.css
@@ -1,0 +1,27 @@
+.held {
+  min-height: 60px;
+  min-width: 60px;
+  opacity: 0;
+  place-self: start;
+}
+
+.transform-held {
+  animation-duration: 1.75s;
+  animation-fill-mode: forwards;
+  animation-name: held-fade-in;
+  animation-timing-function: ease-out;
+}
+
+@keyframes held-fade-in {
+  0% {
+    opacity: 0;
+  }
+  75% {
+    opacity: 0;
+  }
+  100% {
+    border: 5px solid #f3f4f6;
+    border-radius: 0.25rem;
+    opacity: 1;
+  }
+}
