+++ styles/components/game/Queue.module.css
@@ -1,0 +1,32 @@
+.queue {
+  min-width: 120px;
+
+  @apply grid grid-rows-6 gap-1;
+}
+
+.next {
+  min-height: 60px;
+  min-width: 60px;
+  place-self: start;
+}
+
+.transform-queue > .next {
+  animation-duration: 1.75s;
+  animation-fill-mode: forwards;
+  animation-name: queue-fade-in;
+  animation-timing-function: ease-out;
+}
+
+@keyframes queue-fade-in {
+  0% {
+    opacity: 0;
+  }
+  75% {
+    opacity: 0;
+  }
+  100% {
+    border: 5px solid #f3f4f6;
+    border-radius: 0.25rem;
+    opacity: 1;
+  }
+}
