+++ styles/components/game/Top.module.css
@@ -1,0 +1,29 @@
+.info {
+  width: 280px;
+  color: #f3f4f6;
+  font-size: 1.5em;
+  font-weight: 700;
+  opacity: 0;
+  text-align: center;
+
+  @apply grid grid-cols-2 gap-x-3;
+}
+
+.transform-info {
+  animation-duration: 1.75s;
+  animation-fill-mode: forwards;
+  animation-name: info-fade-in;
+  animation-timing-function: ease-out;
+}
+
+@keyframes info-fade-in {
+  0% {
+    opacity: 0;
+  }
+  75% {
+    opacity: 0;
+  }
+  100% {
+    opacity: 1;
+  }
+}
