+++ styles/components/prompt/Control.module.css
@@ -5,7 +5,7 @@
   left: 50%;
   top: 50%;
   transform: translate(-50%, -50%);
-  z-index: 100;
+  z-index: 50;
 
-  @apply grid gap-y-1;
+  @apply grid gap-y-2;
 }
