+++ styles/components/prompt/Countdown.module.css
@@ -4,7 +4,7 @@
   left: 50%;
   top: 50%;
   transform: translate(-50%, -50%);
-  z-index: 90;
+  z-index: 50;
 }
 
 .countdown p {
