+++ typings/tetris.d.ts
@@ -1,0 +1,12 @@
+/* Tetris game state typings */
+declare type TetrisState = {
+  level: number;
+  score: number;
+  field: {
+    colArr: number[];
+    lowestY: number;
+  }[];
+  gameOver: boolean;
+  heldTetromino: TetrominoType;
+  spawnedTetrominos: TetrominoType[];
+};
